+++ src/extensionState.ts
@@ -34,9 +34,11 @@ export const DEFAULT_REPO_STATE: GitRepoState = {
 	pullRequestConfig: null,
 	showRemoteBranches: true,
 	showRemoteBranchesV2: BooleanOverride.Default,
+	simplifyByDecoration: BooleanOverride.Default,
 	showStashes: BooleanOverride.Default,
 	showTags: BooleanOverride.Default,
-	workspaceFolderIndex: null
+	workspaceFolderIndex: null,
+	isCdvSummaryHidden: false
 };
 
 const DEFAULT_GIT_GRAPH_VIEW_GLOBAL_STATE: GitGraphViewGlobalState = {